requirements.txt
fastapi==0.95.2
uvicorn==0.15.0
pandas==1.4.4
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==0.19.0
requests==2.26.0
joblib>=1.1.1
python-multipart==0.0.5
sqlalchemy==1.4.23
pydantic==1.10.7 